chore: refactor object_movement
